test: use puppeteer to test all examples
This commit replaces the itowns-testing.js hack with puppeteer (https://github.com/GoogleChrome/puppeteer).

Advantages:
  - a real browser (Chrome) instead of node + mocks for missing features
  - WebGL capable
  - Can take screenshots and save them (not used on Travis yet)
  - Debuggable using Chrome Dev tools (Chrome instance used by puppeteer can be headless or not)

Each example now have at least one test to verify we can open it and that it
reaches the "mainloop has finished running" state.

I also implemented a couple tests for other examples (see 3dtiles and globe).

Adding test is easy, and we could add more to exercise more parts of itowns.

Last but not least: with this change the separation of examples in 2 files (html + js) is
not needed anymore.